+++ nodejs/awsx/apigateway/index.ts
@@ -13,3 +13,12 @@
 // limitations under the License.
 
 export * from "./api";
+export * from "./authorizer";
+
+// @pulumi/awsx is a deployment-only module.  If someone tries to capture it, and we fail for some
+// reason we want to give a good message about what the problem likely is.  Note that capturing a
+// deployment time module can be ok in some cases.  For example, using "awsx.apigateway.AuthorizerResponse"
+// as a helper function is fine. However, in general, the majority of this API is not safe to use
+// at 'run time' and will fail.
+/* @internal */
+export const deploymentOnlyModule = true;